# [wasm] Jiterpreter v128 misc optimizations