--- a/src/mono/wasm/runtime/jiterpreter-support.ts
+++ b/src/mono/wasm/runtime/jiterpreter-support.ts
@@ -375,6 +375,20 @@ export class WasmBuilder {
         this.appendLeb(value);
     }
 
+    v128_const(value: 0 | Uint8Array) {
+        if (value === 0) {
+            // This encoding is much smaller than a v128_const
+            this.i52_const(0);
+            this.appendSimd(WasmSimdOpcode.i64x2_splat);
+        } else if (typeof (value) === "object") {
+            mono_assert(value.byteLength === 16, "Expected v128_const arg to be 16 bytes in size");
+            this.appendSimd(WasmSimdOpcode.v128_const);
+            this.appendBytes(value);
+        } else {
+            throw new Error("Expected v128_const arg to be 0 or a Uint8Array");
+        }
+    }
+
     defineType(
         name: string, parameters: { [name: string]: WasmValtype }, returnType: WasmValtype,
         permanent: boolean
@@ -1507,6 +1521,19 @@ export function try_append_memset_fast(builder: WasmBuilder, localOffset: number
         builder.local("math_lhs32", WasmOpcode.set_local);
 
     let offset = destOnStack ? 0 : localOffset;
+
+    if (builder.options.enableSimd) {
+        const sizeofV128 = 16;
+        while (count >= sizeofV128) {
+            builder.local(destLocal);
+            builder.v128_const(0);
+            builder.appendSimd(WasmSimdOpcode.v128_store);
+            builder.appendMemarg(offset, 0);
+            offset += sizeofV128;
+            count -= sizeofV128;
+        }
+    }
+
     // Do blocks of 8-byte sets first for smaller/faster code
     while (count >= 8) {
         builder.local(destLocal);
@@ -1586,6 +1613,21 @@ export function try_append_memmove_fast(
 
     let destOffset = addressesOnStack ? 0 : destLocalOffset,
         srcOffset = addressesOnStack ? 0 : srcLocalOffset;
+
+    if (builder.options.enableSimd) {
+        const sizeofV128 = 16;
+        while (count >= sizeofV128) {
+            builder.local(destLocal);
+            builder.local(srcLocal);
+            builder.appendSimd(WasmSimdOpcode.v128_load);
+            builder.appendMemarg(srcOffset, 0);
+            builder.appendSimd(WasmSimdOpcode.v128_store);
+            builder.appendMemarg(destOffset, 0);
+            destOffset += sizeofV128;
+            srcOffset += sizeofV128;
+            count -= sizeofV128;
+        }
+    }
 
     // Do blocks of 8-byte copies first for smaller/faster code
     while (count >= 8) {