* The encoding I used previously for a v128 zero constant in some places (i64.const 0 + i64 -> v128 splat) isn't optimized by v8, so we can't use it. We're stuck using the enormous v128.const opcode instead.
* Use v128 loads/stores if available when unrolling memsets and memmoves. This appears to improve performance for the BCL in some spots by a measurable amount, and at least for the moves it reduces code size too. (The sets will get worse thanks to the incredible constant representation; using a temp local might help maybe?)