@kg 2023-05-19
Also found a bug in the memset unroller (musl style). Some quick before/after timings:
| measurement | before | after |
|-:|-:|-:|
|                    Span, Reverse bytes |     0.0213ms |     0.0139ms |
|                    Span, Reverse chars |     0.0418ms |     0.0256ms |
|                    Span, IndexOf bytes |     0.5535us |     0.3193us |
|                    Span, IndexOf chars |     0.0881ms |     0.0164ms |
|              Span, SequenceEqual bytes |     0.0326ms |     0.0116ms |
|              Span, SequenceEqual chars |     0.0651ms |     0.0232ms |